refactor: move recaptcha code in preparation
In preparation for a conversion to a proper service, move the recaptcha
code to a package named `captcha`.
Next refactoring will include mass rename of variables.
