+++ tests/unit/test_config.py
@@ -379,7 +379,7 @@ def test_configure(monkeypatch, settings, environment):
             pretend.call(".sentry"),
             pretend.call(".csp"),
             pretend.call(".referrer_policy"),
-            pretend.call(".recaptcha"),
+            pretend.call(".captcha"),
             pretend.call(".http"),
         ]
         + [pretend.call(x) for x in [configurator_settings.get("warehouse.theme")] if x]
